+++ app/javascript/frontend/partials/Support/index.vue
@@ -56,7 +56,3 @@ export default {
   },
 }
 </script>
-
-<style lang="scss" scoped>
-@import 'styles/index.scss';
-</style>