+++ app/javascript/packs/frontend.js
@@ -1,7 +1,6 @@
 import 'stylesheets/frontend'
 import Vue from 'vue'
 import VTooltip from 'v-tooltip'
-import VueClipboard from 'vue-clipboard2'
 import App from 'frontend/App'
 import router from 'frontend/lib/Router'
 import store from 'frontend/lib/Store'
@@ -11,7 +10,6 @@ import 'frontend/lib/Bootstrap'
 import I18nPlugin from 'frontend/lib/I18n'
 import ApiClient from 'frontend/lib/ApiClient'
 import Subscriptions from 'frontend/lib/Subscriptions'
-import VueScrollTo from 'vue-scrollto'
 import Comlink from 'frontend/lib/Comlink'
 import Meta from 'vue-meta'
 import DataPrefetch from 'frontend/lib/DataPrefetch'
@@ -19,22 +17,6 @@ import Helpers from 'frontend/lib/Helpers'
 import Noty from 'frontend/lib/Noty'
 import Ahoy from 'frontend/lib/Ahoy'
 import Validations from 'frontend/lib/Validations'
-
-console.info(`
-
-███████████  ██                             ██     ██      ██                                ██                                          ██
-██           ██                             ██      ██    ██                                 ██                                          ██
-██           ██     ██████████  ██████████  █████    ██  ██    ██████████  ████████  ██████████  ██████████      ██████████  ██████████  █████
-████████     ██     ██      ██  ██      ██  ██        ████             ██  ██        ██      ██  ██              ██      ██  ██      ██  ██
-██           ██     ██████████  ██████████  ██         ██      ██████████  ██        ██      ██  ██████████      ██      ██  ██████████  ██
-██           ██     ██          ██          ██         ██      ██      ██  ██        ██      ██          ██      ██      ██  ██          ██
-██           ████   ██████████  ██████████  █████      ██      ██████████  ██        ██████████  ██████████  ██  ██      ██  ██████████  █████
-
-API: https://api.fleetyards.net
-GITHUB: https://github.com/fleetyards
-TWITTER: https://twitter.com/FleetYardsNet
-
-`)
 
 Vue.use(Subscriptions)
 Vue.use(ApiClient)
@@ -45,14 +27,6 @@ Vue.use(Helpers)
 Vue.use(Noty)
 Vue.use(Ahoy)
 Vue.use(Validations)
-
-Vue.use(VueClipboard)
-
-Vue.use(VueScrollTo, {
-  easing: 'ease-in-out',
-  duration: 1000,
-  offset: -100,
-})
 
 Vue.use(Meta, {
   keyName: 'head',
@@ -79,10 +53,6 @@ Vue.use(VTooltip)
 //   }
 // }
 
-console.info(`APP Version: ${window.APP_VERSION} (${window.APP_CODENAME})`)
-console.info(`Store Version: ${window.STORE_VERSION}`)
-console.info(`API Endpoint: ${window.API_ENDPOINT}`)
-
 document.addEventListener('DOMContentLoaded', () => {
   if ('serviceWorker' in navigator) {
     // eslint-disable-next-line compat/compat
@@ -107,8 +77,6 @@ document.addEventListener('DOMContentLoaded', () => {
     store.dispatch('reset')
     store.commit('setStoreVersion', window.STORE_VERSION)
   }
-
-  store.dispatch('sentry/reset')
 
   // eslint-disable-next-line no-new
   new Vue({