+++ config/webpack/loaders/typescript.js
@@ -2,10 +2,12 @@ const PnpWebpackPlugin = require('pnp-webpack-plugin')
 
 module.exports = {
   test: /\.(ts|tsx)?(\.erb)?$/,
-  use: [{
-    loader: 'ts-loader',
-    options: PnpWebpackPlugin.tsLoaderOptions({
-      appendTsSuffixTo: [/\.vue$/],
-    }),
-  }],
+  use: [
+    {
+      loader: 'ts-loader',
+      options: PnpWebpackPlugin.tsLoaderOptions({
+        appendTsSuffixTo: [/\.vue$/],
+      }),
+    },
+  ],
 }